+++ SacStateApp/src/SacStateColors/LoginColors.js
@@ -1,0 +1,16 @@
+
+//Variable names and colors used in LoginStyles
+const brightRed = '#ff4d4d';
+const darkGreen = '#043927';
+const darkOverlay = 'rgba(0, 0, 0, 0.5)';
+const darkerOverlay = 'rgba(0, 0, 0, 0.6)';
+const fadedGold = 'rgba(251, 248, 239, 0.8)';
+const mutedDarkGreen = 'rgba(43, 58, 50, 0.8)';
+const mutedGold = '#E4CFA3';
+const mutedSacGreen = 'rgba(4, 57, 39, 0.8)';
+const mutedSacStateGreen = 'rgba(4, 57, 39, 0.9)';
+const opaqueWhite = 'rgba(255, 255, 255, 0.9)';
+const softCream = '#FDF6E4';
+const softOverlay = 'rgba(0, 0, 0, 0.4)';
+const subtleFadedGold = 'rgba(251, 248, 239, 0.5)';
+
