+++ SacStateApp/src/SacStateColors/ProfileColors.js
@@ -1,0 +1,5 @@
+
+const fadedSacGold = '#FBF8EF';
+const mutedGold = '#E4CFA3';
+const sacGreen = '#043927';
+const whiteSmoke = '#F5F5F5';
