+++ SacStateApp/src/SacStateColors/ProfileCreationColors.js
@@ -1,0 +1,6 @@
+const fadedSacGold = '#FBF8EF';
+const mutedGold = '#E4CFA3';
+const offWhite = '#F8F9FA';
+const sacGreen = '#043927';
+const subtleGreen = '#E4F1E7';
+const cleanWhite = '#FFFFFF';
