+++ SacStateApp/src/SacStateColors/SettingsColors.js
@@ -1,0 +1,2 @@
+const white = '#fff';
+const sacGreen = '#043927';
