+++ SacStateApp/src/SacStateColors/WellnessColors.js
@@ -1,0 +1,2 @@
+const cleanWhite = '#FFFFFF';
+const sacGreen = '#043927';
