+++ SacStateApp/src/SettingsStyles/SettingsStyles.js
@@ -1,0 +1,17 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#fff', // Background color
+  },
+  text: {
+    fontSize: 24,
+    fontWeight: 'bold',
+    color: '#043927', // Sac State Green
+  },
+});
+
+export default styles;
