+++ SacStateApp/src/screens/SettingsScreen.js
@@ -2,6 +2,7 @@
 
 import React, { useState } from 'react';
 import { View, Text, StyleSheet } from 'react-native';
+import styles from '../SettingsStyles/SettingsStyles'; //new
 
 const SettingsScreen = () => {
   return (
@@ -10,19 +11,5 @@ const SettingsScreen = () => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#fff', // Background color
-  },
-  text: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    color: '#043927', // Sac State Green
-  },
-});
 
 export default SettingsScreen;
