+++ SacStateApp/src/screens/WellnessScreen.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { View, Text, StyleSheet } from 'react-native';
+import styles from '../WellnessStyles/WellnessStyles'; //new
 
 const WellnessScreen = () => {
   return (
@@ -8,19 +9,5 @@ const WellnessScreen = () => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#FFFFFF', // White background
-  },
-  text: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    color: '#043927', // Sac State green
-  },
-});
 
 export default WellnessScreen;
